Move all views related to Lane to a single file called LaneView.
